requirements.txt
flask
flask-admin
flask-babelex
email_validator
Flask-SQLAlchemy

Flask-WTF
flask_bootstrap

 # beacuse imports in pypi broken -https://github.com/maxcountryman/flask-uploads/issues/29
git+https://github.com/maxcountryman/flask-uploads.git


qrcode[pil]

phonenumbers

flask-login